rename FavoriCard to FavoriteCard
