+++ web-app/src/components/settings/FavoriteCard.tsx
